+++ piano-analytics/src/main/java/io/piano/android/analytics/EventRepository.kt
@@ -22,6 +22,15 @@ internal class EventRepository(
         )
     }
 
+    fun deleteOutOfLimitNotSentEvents(limit: Int) {
+        databaseHelper.delete(
+            "${EventRecord.IS_SENT} = 0 AND ${EventRecord.ID} <= (" +
+                "SELECT ${EventRecord.ID} FROM ${EventRecord.TABLE_NAME} WHERE ${EventRecord.IS_SENT} = 0 " +
+                "ORDER by ${EventRecord.ID} DESC LIMIT 1 OFFSET ?)",
+            limit.toString()
+        )
+    }
+
     fun getNotSentEvents(): List<EventRecord> = databaseHelper.query(
         selection = "${EventRecord.IS_SENT} = 0",
         orderBy = "${EventRecord.TIME} ASC"