+++ piano-analytics/src/main/java/io/piano/android/analytics/SendTask.kt
@@ -24,6 +24,7 @@ internal class SendTask(
 ) : Runnable {
     override fun run() {
         eventRepository.deleteOldEvents(configuration.eventsOfflineStorageLifetime)
+        eventRepository.deleteOutOfLimitNotSentEvents(EVENTS_LIMIT)
         if (deviceInfoProvider.connectionType == ConnectionType.OFFLINE) {
             Timber.w("Can't send events - no connection")
             return
@@ -61,5 +62,6 @@ internal class SendTask(
     companion object {
         internal val MEDIA_TYPE by lazy(LazyThreadSafetyMode.NONE) { "application/json; charset=UTF-8".toMediaType() }
         private const val CHUNK_SIZE = 50
+        private const val EVENTS_LIMIT = 2000
     }
 }